vznetaddbr: fix interface initialization
Initially, we used "ifconfig $if 0" to initialize veth interface.

Commit 93ef4d978 changed that to using "ip a a 0 dev $ifname" which
looks like full analog of previous ifconfig, but lead to bug #1089
reported.

Next, commit 77631ba55 (ironically with the same same subject as this one)
changed "ip a a 0 dev $if" to "ip link set dev $if up". This probably
fixed bug #1089, but lead to opening a few others: #1617, #1731, #1682.

So this is yet another attempt to fix it. Now we use both:

	ip link set dev $if up
	ip a a 0 dev $if

If this won't work for everyone, I will just revert to using ifconfig.

http://bugzilla.openvz.org/1617

Signed-off-by: Kir Kolyshkin <kir@openvz.org>