etc/dists/scripts/debian-add_ip.sh: remove all traces of venet0
When starting a container with no IP addresses configured, make
sure to remove everything about venet0 from /etc/network/interfaces.

Before this patch, we only checked if there are aliases
(like venet0:1), but not for venet0 itself. Now we check
both for venet0 and its aliases.

This is part of the fix for bug #1596.

http://bugzilla.openvz.org/1596

Reported-by: Olena Vasilkova <daemon@root.ua>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>