vzctl restore: CPT_UNDUMP may return positive value
When vzctl writes something to CPT_LOCKFD, the kernel reads data
from this file descriptor and should return an error (negative value),
but instead it returns code from read() (naturally, positive value).

This kernel bug exists in old kernels (every one except rhel6 < 042test008).

Indeed, this is the bug in the kernel, so this patch is a workaround for it.
Instead of checking for negative value, we check for non-zero value,
which is also OK.

http://bugzilla.openvz.org/1561

[kir@openvz.org: improved description]

Signed-off-by: Andrey Vagin <avagin@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>