vzctl create: exit with error if default CT config not found
It was discovered that vzctl create tolerates invalid setting of CONFIGFILE
in /etc/vz/vz.conf and just silently creates a container with unusable
configuration file.

Fix it by explicitly checking that config sample exists, and error out
if it's not, giving a user a helpful hint of what's wrong, like this:

 # vzctl create 21342134
 File /etc/vz/conf/ve-something.conf-sample not found: No such file or directory
 Fix the value of CONFIGFILE in /etc/vz/vz.conf
 Creation of container private area failed

Note that we are giving the error even if we are not going to use
default config (such as when CT config already exists).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>