etc/dists/scripts/debian-add_ip.sh: fix not adding default route
This fixes the problem of non-working network in Debian and Ubuntu
containers after upgrade to vzctl-3.0.26, caused by commit 9612a0121.

Apparently, older kernels (like 2.6.18) are refusing to add default
route for a device with no IP addresses. Such problem does not exist
with RHEL6 based (2.6.32) kernel, but RHEL5-based kernels (2.6.18)
are failing miserably.

The workaround is to use 'ifconfig venet0 0' which apparently adds
something like 0.0.0.0/0 IP to venet0, after which "route add" works.

The issue is reported in bug #1792:
http://bugzilla.openvz.org/1792

It is also referred to in bug #1669 (comment #6):
http://bugzilla.openvz.org/1669

[kir@openvz.org: changelog entry]