etc/init.d/vz*: more proper kernel check in modify_vzconf()
For now, the only VSwap-capable kernel is 042{test,stab}.
Currently all the kernels that do not have stab or test in their name
(ie in /proc/vz/version) are definitely not VSwap-capable. If they
ever will be we will need to modify this check.

Add the check for 'test|stab' to avoid the following error while
starting vz initscript:

  /etc/init.d/vz: line 163: test: dzhanibekov: integer expression expected

Reported-by: Stanislav Kinsbursky <skinsbursky@openvz.org>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>