Fix networking for Debian-3.1
After a few changes in dists/scripts (specifically, 9612a012)
we broke the compatibility with Debian 3.1. My guess was nobody
will notice since people have moved on to newer versions, but
apparently I was wrong.

Let's fix it in a very simple way -- use older version of
add_ip script for Debian-3.x, generated with

  git show 25c14e96:debian-add_ip.sh

Note that del_ip script is about the same so we use the same one
for all Debian versions

http://bugzilla.openvz.org/1802

Signed-off-by: Kir Kolyshkin <kir@openvz.org>