vznetaddbr: fix interface initialization (again and again)
This is yet another attempt to fix veth initialization after we
switched from ifconfig to ip (apparently breaking the
"don't touch it if it works" rule).

For historical retrospective of this issue, see the following

* git commits: 3c638947, 77631ba55, 93ef4d978.
* bug reports: #1617, #1731, #1682, #1089.

This time, the problem is some intermediate versions of 'ip' utility
were broken in 2008 to not understand the abbreviated IP address
('0' as '0.0.0.0/0'). For example, see this discussion:

http://www.kerneltrap.org/mailarchive/linux-netdev/2008/10/8/3556034/thread

Apparently some Linux distros do have such broken ip versions.
So, in order for these to work, let's not abbreviate IP and use
full 0.0.0.0/0 instead of just 0. Hope it will fix bug #1831.

PS as promised in commit 3c638947, next time I will just revert to
using ifconfig which was always working.

http://bugzilla.openvz.org/1831

Signed-off-by: Kir Kolyshkin <kir@openvz.org>
(cherry picked from commit 057b0d7fc600996ff87353357eae6818770558bb)