vzgetnetdev: filter out link-local addresses
When we make the list of NETDEVICES (to be used to send ARP/NDP
announces/queries to) we try hard to remove all the "bad" devices,
such as ones that are NOARP, SLAVE, doesn't have any IPs assigned
etc.

This patch adds one more such filter: if an interface only has
link-local addresses, it should not be used.

http://bugzilla.openvz.org/1820

Reported-by: Ekaterina Epifanova <nixma@yandex.ru>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>