Fix: --quiet ignored for global config options parsing
Here is the bug:

	# vzctl --quiet start 202
	Warning: Unknown iptable module: ipt_iprange, skipped
	Warning: Unknown iptable module: sinexiud, skipped

Apparently --quiet is ignored here. The reason is init_log() is called
two times (second time after parsing global config, so the values of
LOG_FILE, VERBOSE etc are known), and for the first time it is called
without passing the quiet flag.

Fix is easy.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>