vzctl: do report unknown parameters in config
When we are parsing configuration files, unknown parameters are ignored,
and this behaviour is documented at least in vz.conf(5) and ctid.conf{5}
man pages. It makes sense because it is possible to add some additional
parameters to config which are to be used by tools other than vzctl.

Nevertheless, sometimes it makes sense to report unknown parameters
(for example, if there is a typo).

This commit makes vzctl report such unrecognized parameters, but it does
that with a log level of 1, so it is not shown by default. In order to see
it, one have to increase log level or verbosity (for example, by running
vzctl --verbose).

While at it, add a comment in parse_opt() that ERR_UNK is not expected here.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>