vzubc: fix controversion between CTID and file name
Now to specify a file name (which is very rarely needed, mostly for debug)
one needs to use -f option. So the only use for 'non-option' arguments is
CTID or CT name.

Thus, it is now possible to use CTIDs of deleted containers (when a container
is stopped/removed, but beancounters still remain).

Fix usage() and the man page accordingly.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>