vzctl set --quotaugidlimit: warn on turning quota on/off, no warn on change
When a value of quotaugidlimit changes, there's no need to warn, it can
be done on the fly. But when we turn the quota on or off -- we need to warn.
So fix check_set_mode() accordingly.

In fact we can't catch all the cases here, since we'd need to check the
actual quota state as well (and it is done later in the code), but let's
try to catch what we can catch.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>