man/vzlist.8: reformat the tables
Make the table with possible fields smaller, by removing the
versions with .m .b .l .f suffixes. This way it fits in 1 page.

Split the table with suffixes to two tables: for beancounters
and for disk quotas. Move this after the fields table.

Fix formatting in example, because text-based man do not understand \f(CW.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>