vzctl set: add ability to parse numbers like 1.5G
This adds the ability to specify values using floating point
numbers. Surely this only makes sense with a suffix, such as
1.5G or 0.5T.

Note that we try to parse the number as long long first, this kinda
provides backward compatibility.

Also, since we are using long long double in internal calculations,
we can now detect overflow better. Sorry I do not remember an example
of that at the moment.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>