vzubc: add support for reading /proc/bc/$ID/resources
It was not possible to use it before, because we relied on BCID
field to be present. In /proc/bc/resources and /proc/user_beancounters
it is printed just before 'kmemsize', but in /proc/bc/*/resources
it is nowhere to be found.

Change the logic to not rely on BCID to be known.

We actually need BCID for relative and incremental modes, so
add a check that BCID is there, and error out if it is not.

With this patch, it is now possible to do things like:

	vzubc -f /proc/bc/101/resources

or even

	cat /proc/bc/{105,106}/resources | vzubc -f -

but do not expect -r|-i to work. Best way to do such things is just:

	vzubc 105 106

Signed-off-by: Kir Kolyshkin <kir@openvz.org>