scripts/vzevent-stop: try to not get in the way of vzctl stop
If vzctl stop is running for a container, it stops then unmounts
the container. Event 've-stop' is delivered in the middle of the
'vzctl stop' run so vzevent-stop is executed while 'vzctl stop'
is still running. Same logic applies to 'vzctl restart'.

If 've-stop' event is caused by 'vzctl stop' or 'vzctl restart'
we don't want to perform any actions in vzevent-stop script.
So, if there is such a process running, just exit silently and
let it do the job.

This probably fixes bug #1878
http://bugzilla.openvz.org/1878

Reported-by: Dietmar Maurer <dietmar@maurer-it.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>