Be less chatty when running distscripts
Logging 'File resolv.conf was modified' with a normal log level is
pointless. First, it might not be resolv.conf (what it really is is
up to set_dns script). Second, a script might have returned an error
but we still say "was modified". Third, normal log level is too high.
Fourth, we might reduce the level to debug, so it would only be seen
with --verbose, but in that case we will anyway see the message from
vps_exec_script ("Running container script xxx").

Similar concerns for the other two messages.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>