set_ugid_quota.sh: fix filesystem type
Commit e324b0ad changed filesystem type in created /etc/mtab from reiserfs
to simfs:

> 4. Remove old/questionable reiserfs reference. No one knowns why it was ther
>    (reiserfs quota tools? who knows).

Now we know why it was there. Below is an essential, see the long
description in http://bugzilla.openvz.org/1904#c31

(a) Some of the quota utilities have a build-in list of recognized
filesystems, and refuse to work with others. Therefore we have to use some
known name.

(b) quotacheck has some extra assumptions/knowledge about ext3 and fails
in CT, while reiserfs is known to not cause such problems.

Revert back to good old 'reiserfs'.

This should fix bug #1904 for older quota tools (i.e. ones that read
/etc/mtab). The rest (i.e. /proc/mounts) is to be fixed in the kernel.

http://bugzilla.openvz.org/1904

Reported-by: John Criss <admin@nginxcp.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>