vzctl set --ipadd: treat IPs with different masks as the same one
Introduce merge_ip_list() which is the same as merge_str_list(),
only using find_ip() instead of find_str() for checking IP addresses.
Since find_ip tries not to look at netmask (commit 0f8c65271), we
will not have same IPs with different masks in the resulting list.

Also, due to commit e3d0346df new values have precedence over the
old ones, so --ipadd can be used to change the mask of the already
existing IP. Of course this is up to etc/dists/scripts/*_ip.sh
to process this in a correct fashion.

http://bugzilla.openvz.org/1088

Signed-off-by: Kir Kolyshkin <kir@openvz.org>