__merge_str_list(): 'add' list have a precedence over 'del'
Consider the following command:

	vzctl set --ipdel 1.2.3.4 --ipadd 1.2.3.4

Before this patch, this does nothing. Now, we are adding values from
the 'add' list even if they are in the 'del' list, thus forcing the
command to remove the IP and add it again. This totally makes sense,
provided we remove IPs first and then add those (need to check it).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>