vzctl set --ipadd: fix for adding existing IP with different netmask
If we are adding an IP, and the same IP is already in config (possibly
with different netmask), add the IP from config to 'del' list (as
if a user specified --ipdel). It will cause the from-config IP to be deleted
and the from-command-line IP to be added. This makes most sense if netmask
has changed, but even if it is the same, that kinda means "re-add this IP".

This patch relies on previous commit 3249c0d42.

http://bugzilla.openvz.org/1088

Signed-off-by: Kir Kolyshkin <kir@openvz.org>