vzlist: fix segfault when sorting by loadavg
Apparently the sorting (1) didn't work and (2) segfaulted in case some CTs didn't have
load average values (which is pretty easy to have if you list stopped CTs).

BTW this is a zero day bug (it comes from the initial vzctl commit to git), so it took
about 6 years to discover it. My congratulations to Alexander!

Reported-by: Shvayakov Alexander <a.shvayakov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>