etc/dists/scripts/debian-del_ip.sh: fix removing too much IPv6 addresses
Commit 7b139a9e87 introduced IPv6 support for Debian. IPv6 address removal
was performed like this:

	grep -v $ip $${CFGFILE}

So, if you want to remove an address like 10:29::1, addresses like
10:29::10, 10:29::145 etc. will also be removed (just checked, it
indeed happens). This is definitely not what we want here.

Adding -w helps. Alternatively, we can add a space before and a slash
after, but -w works good enough.

While at it, add -F since we do look for a fixed string not a regex.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>