etc/dists/scripts/debian-del_ip.sh: try harder to remove all IPs
Two improvements:

(1) Try to down venet0 aliases before downing the venet0 itself. This results
in no venet0:X left (in down state but with an IP address).

(2) Use 'ifconfig $dev down' instead of 'ifdown $dev' because the latter
needs a configuration in /etc/network/interfaces while the former doesn't
(there might be some corner cases then we want to remove an address but
it has no record in /etc/network/interfaces).

The only address not removed after this is 127.0.0.2. I guess it's harmless
(and the interface is down anyway).

Signed-off-by: Kir Kolyshkin <kir@openvz.org>