vzctl set: tailor "no save" warning to running/stopped states
Before this patch, the warning printed in case no --save
flag is supplied to vzctl set command looked like this:

WARNING: Settings were not saved and will be reset to original values at the next start (use --save flag)

Obviously, the 'and will be reset to original values at the next start'
part only makes sense if the container is running. So make this part
conditional.

While at it, rephrase the warning a bit.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>