cap.c: if capset fails, retry with old default
Apparently RHEL5 kernel doesn't like CAP_SETPCAP being set.
It was added to default set in commit 4721d0e68e and released
in vzctl-3.0.30, which doesn't work with RHEL5 kernel (CT is
not started).

The solution is to silently retry without this capability, assuming
the problem is the old kernel.

While at it
* remove SETFCAP (added by commit f319aef1d) from the "old" set
* untangle the brackets madness messed up in commit f319aef1d

Signed-off-by: Kir Kolyshkin <kir@openvz.org>