etc/dists/scripts/gentoo-add_ip.sh: do not set up venet0 if no IPs
If you configure a CT to only use veth interfaces (no IP addresses are
configured for venet), gentoo-add_ip.sh still adds net.venet0 to the
defalt runlevel and sets default route via venet0.

Solution is not to run init_netconfig() on CT start if there are no IPs,
but it creates a problem of adding first IP address to a running container --
in this case routing for venet0 is not configured. This problem is fixed
by using "have_ips && venet_configured || set_config" for running CT.

http://bugzilla.openvz.org/2077

Reported-by: Frank Wassmuth <frank@wassmuth.it>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>