etc/init.d/vz: Stricter auto-replacement of CONFIGFILE
Commits 39bdf8e, c5995be, 75678d added a feature to switch CONFIGFILE
from default (basic) to vswap-enabled one, if CONFIGFILE was not touched
(i.e. same as in shipped configuration file).

Apparently this was bugging some customers who want to use
CONFIGFILE="basic".

To solve this, add a comment to config file, and include that comment
into check.

Now, in order for this autoreplacement to not work, remove that comment.

http://bugzilla.openvz.org/2169

Reported-by: Pavel Odintsov <pavel.odintsov@gmail.com>
Signed-off-by: Kir Kolyshkin <kir@openvz.org>