ploop: etc/init.d/vz-*: load/unload ploop modules
They are only available in latest RHEL6-based kernel, but since we are
ignoring modprobe errors it's fine to try loading/unloading these modules
even if there are no modules. In other words, there could be kernel version
checks added but from my perspective it's OK to go without it.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>