ploop: introduce VE_LAYOUT / --layout
Now one can choose whether to put a new CT on ploop or use old simfs.

First, there is a global VE_LAYOUT parameter in vz.conf.
Second, there is a --layout option for vzctl create.
Default for now is simfs (might be changed in the future).

If ploop is requested, a check is performed if ploop is supported by the kernel.

Man pages and bash_completion modified accordingly.

(this is an implementation of TODO1 mentioned in commit 27cea49)

Signed-off-by: Kir Kolyshkin <kir@openvz.org>