ploop: resize ploop image when vzctl set --diskspace is run
Previous commit ebc46fc was incorrect, because vps_setup_res() is run
for every vzctl start and, if a container is running, for vzctl set.
This is the right place for vzquota, but not for ploop.

For ploop we only need to do it on vzctl set stage, never minding the
container status, and we don't have to do anything on vzctl start.

Move the ploop resizing call to set().

Signed-off-by: Kir Kolyshkin <kir@openvz.org>