vzctl set: remove --noatime
The --noatime flag was added many years ago (in 2005 or earlier) in order to
work around performance degradation caused by need to write down new atime
after every file access.

This problem is currently solved by relatime feature which is there at least
in RHEL5 and RHEL6 kernel and is enabled by default. So, --noatime is not
making much sense now, and we can safely drop it.

To not introduce incompatibility, the flag, if given from the command line,
is ignored, as well as NOATIME config parameter.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>