Move OVZ-specific vz_open() to hooks_vz.c
From now on, we'll assume that if there is no /dev/vzctl, we're not
running on top of OVZ kernel, but rather, on normal upstream linux.

If that is the case, we'll refrain to do any VZ-specific operation, and
rely on what we can do from upstream linux.

A new file is added, lib/hooks_vz.c. All the VZ-specific functionality
will be gradually moved there. A corresponding upstream linux function
will follow.

[kir@openvz.org: changelog and other minor fixes]
Signed-off-by: Glauber Costa <glommer@parallels.com>